Remove JS build environment
The JavaScript build environment did in reality just two things:
it added Skeleton, and it minified the HTML and CSS. Add Skeleton
as a part of the repository (it is also licensed under MIT), and
add instructions recommending caching and gzip-ing (greater
size reduction would be gained by scaling down the greatly oversized
images).

The reason for not wanting a full-stack JavaScript environment is that
its impact on readability of the code and modifiability (due to simplifying
inclusion of libraries and compilation from less to CSS) does not weight
up for the hit on modifiability (due to the complexity of setting this
up, especially if you're not developing JavaScript apps day by day) and
ease of deployment (requiring you to basically have a development
environment just to push out a small change). This web page is very simple,
the use of less had no impact whatsoever, and the use of JavaScript Webpack,
bower etc. complicates things much more than they help. If the web
page used React, was interactive, greatly benefitted from using less
or something along those lines, the JavaScript stack would be much more appropriate.