Ensure names are not split into two lines
"Under Dusken" and "Radio Revolt" look weird when split up so one
word appears on one line, and the other word on the next line.
This commit makes it so they are not broken up, due to
Non Breakable SPace. They are not long enough for this to be much
of an issue.