Add web publishing services to video section
The web publishing tools are not exclusively for writers,
but are also for video creators. Labelling them as writer's tools
only could make video creators feel excluded. Therefore, list
them in the video section as well.

As an unfortunate side effect, the number of services in
the video section must be four, not three like everywhere else.
I think this is acceptable because one of the tools, Chimera,
will be removed in the not too distant future. I don't think
adding one more row would outweight pushing the footer below
the fold.